requirements.txt
yt-dlp
python-mpv
